+++ doc/changelog.rst
@@ -55,6 +55,7 @@ Addons
 Front
 ~~~~~
 
+- Checkout: Fix method phase attribute population
 - Send registration activation e-mail via notify event
 - Cusmoter information: Replace untranslated "Not specified" with a dash
 - Trigger shipment created event when addons have already processed it